requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-import-export==2.4.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
pytz==2020.5
PyYAML==5.3.1
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
graphene-django==2.15.0
django-cors-headers==3.7.0